Refactor and trace.
