+++ src/Tests/SenseNet.BlobStorage.IntegrationTests/BuiltInProviderTests.cs
@@ -16,19 +16,6 @@ namespace SenseNet.BlobStorage.IntegrationTests
         protected override Type ExpectedExternalBlobProviderType => null;
         protected override Type ExpectedMetadataProviderType => typeof(MsSqlBlobMetaDataProvider);
 
-        protected override void BuildLegoBricks(RepositoryBuilder builder)
-        {
-            Configuration.BlobStorage.BlobProviderClassName = ExpectedExternalBlobProviderType?.FullName;
-            BuiltInBlobProviderSelector.ExternalBlobProvider = null; // reset external provider
-
-            var blobMetaDataProvider = (IBlobStorageMetaDataProvider)Activator.CreateInstance(ExpectedMetadataProviderType);
-
-            builder
-                .UseBlobMetaDataProvider(blobMetaDataProvider)
-                .UseBlobProviderSelector(new BuiltInBlobProviderSelector());
-
-            BlobStorageComponents.DataProvider = blobMetaDataProvider;
-        }
         protected internal override void ConfigureMinimumSizeForFileStreamInBytes(int newValue, out int oldValue)
         {
             // do nothing
